requirements.txt
kopf==1.35.4
kubernetes==23.3.0
PyYAML==6.0
